+++ lib/src/material/date_utils.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/material.dart';
+import 'package:nepali_date_picker/nepali_date_picker.dart';
 import 'package:nepali_utils/nepali_utils.dart';
 
 /// Returns a [NepaliDateTime] with just the date of the original, but no time set.
@@ -47,4 +49,38 @@ int firstDayOffset(int year, int month) {
 /// Returns the number of days in a month.
 int getDaysInMonth(int year, int month) {
   return NepaliDateTime(year, month).totalDays;
+}
+
+/// Returns a locale-appropriate string to describe the start of a date range.
+///
+/// If `startDate` is null, then it defaults to 'Start Date', otherwise if it
+/// is in the same year as the `endDate` then it will use the short month
+/// day format (i.e. 'Asr 21'). Otherwise it will return the short date format
+/// (i.e. 'Asr 21, 2077').
+String formatRangeStartDate(MaterialLocalizations localizations, NepaliDateTime startDate, NepaliDateTime endDate) {
+  return startDate == null
+      ? localizations.dateRangeStartLabel
+      : (endDate == null || startDate.year == endDate.year)
+          ? NepaliDateFormat('MMMM d').format(startDate)
+          : NepaliDateFormat.yMd().format(startDate);
+}
+
+/// Returns an locale-appropriate string to describe the end of a date range.
+///
+/// If `endDate` is null, then it defaults to 'End Date', otherwise if it
+/// is in the same year as the `startDate` and the `currentDate` then it will
+/// just use the short month day format (i.e. 'Asr 21'), otherwise it will
+/// include the year (i.e. 'Asr 21, 2077').
+String formatRangeEndDate(
+    MaterialLocalizations localizations, NepaliDateTime startDate, NepaliDateTime endDate, NepaliDateTime currentDate) {
+  return endDate == null
+      ? localizations.dateRangeEndLabel
+      : (startDate != null && startDate.year == endDate.year && startDate.year == currentDate.year)
+          ? NepaliDateFormat('MMMM d').format(endDate)
+          : NepaliDateFormat.yMd().format(endDate);
+}
+
+/// Returns a [NepaliDateTimeRange] with the dates of the original without any times set.
+NepaliDateTimeRange datesOnly(NepaliDateTimeRange range) {
+  return NepaliDateTimeRange(start: dateOnly(range.start), end: dateOnly(range.end));
 }